pkp/pkp-lib#1604 restructuring for doaj plugin
