+++ classes/article/PublishedArticleDAO.inc.php
@@ -733,9 +733,9 @@ class PublishedArticleDAO extends ArticleDAO {
 
 
 	/**
-	 * Get all published submissions with a pubId assigned and matching the specified settings.
-	 * @param $pubIdType string
+	 * Get all published submissions (eventually with a pubId assigned and) matching the specified settings.
 	 * @param $contextId integer optional
+	 * @param $pubIdType string
 	 * @param $title string optional
 	 * @param $author string optional
 	 * @param $issueId integer optional
@@ -745,15 +745,15 @@ class PublishedArticleDAO extends ArticleDAO {
 	 * @param $rangeInfo DBResultRange optional
 	 * @return DAOResultFactory
 	 */
-	function getByPubIdType($pubIdType, $contextId = null, $title = null, $author = null, $issueId = null, $pubIdSettingName = null, $pubIdSettingValue = null, $rangeInfo = null) {
+	function getExportable($contextId, $pubIdType = null, $title = null, $author = null, $issueId = null, $pubIdSettingName = null, $pubIdSettingValue = null, $rangeInfo = null) {
 		$params = array();
 		if ($pubIdSettingName) {
 			$params[] = $pubIdSettingName;
 		}
 		$params = array_merge($params, $this->getFetchParameters()); // because of the neccessary section row names in _fromRow
-		$params[] = 'pub-id::'.$pubIdType;
-		if ($contextId) {
-			$params[] = (int) $contextId;
+		$params[] = (int) $contextId;
+		if ($pubIdType) {
+			$params[] = 'pub-id::'.$pubIdType;
 		}
 		if ($title) {
 			$params[] = 'title';
@@ -764,8 +764,8 @@ class PublishedArticleDAO extends ArticleDAO {
 		if ($issueId) {
 			$params[] = (int) $issueId;
 		}
-		import('classes.plugins.DOIPubIdExportPlugin');
-		if ($pubIdSettingName && $pubIdSettingValue && $pubIdSettingValue != DOI_EXPORT_STATUS_NOT_DEPOSITED) {
+		import('classes.plugins.PubObjectsExportPlugin');
+		if ($pubIdSettingName && $pubIdSettingValue && $pubIdSettingValue != EXPORT_STATUS_NOT_DEPOSITED) {
 			$params[] = $pubIdSettingValue;
 		}
 
@@ -775,19 +775,19 @@ class PublishedArticleDAO extends ArticleDAO {
 			FROM	published_submissions ps
 				JOIN issues i ON (ps.issue_id = i.issue_id)
 				LEFT JOIN submissions s ON (s.submission_id = ps.submission_id)
-				LEFT JOIN submission_settings ss ON (s.submission_id = ss.submission_id)
-				' . ($title != null?' LEFT JOIN submission_settings sst ON (s.submission_id = sst.submission_id)':'')
+				' . ($pubIdType != null?' LEFT JOIN submission_settings ss ON (s.submission_id = ss.submission_id)':'')
+				. ($title != null?' LEFT JOIN submission_settings sst ON (s.submission_id = sst.submission_id)':'')
 				. ($author != null?' LEFT JOIN authors au ON (s.submission_id = au.submission_id)':'')
 				. ($pubIdSettingName != null?' LEFT JOIN submission_settings sss ON (s.submission_id = sss.submission_id AND sss.setting_name = ?)':'')
 				. ' ' . $this->getFetchJoins() .'
 			WHERE
-				i.published = 1 AND ss.setting_name = ? AND ss.setting_value IS NOT NULL
-				' . ($contextId != null?' AND s.context_id = ?':'')
+				i.published = 1 AND s.context_id = ?
+				' . ($pubIdType != null?' AND ss.setting_name = ? AND ss.setting_value IS NOT NULL':'')
 				. ($title != null?' AND (sst.setting_name = ? AND sst.locale = ? AND sst.setting_value LIKE ?)':'')
 				. ($author != null?' AND (au.first_name LIKE ? OR au.middle_name LIKE ? OR au.last_name LIKE ?)':'')
 				. ($issueId != null?' AND ps.issue_id = ?':'')
-				. (($pubIdSettingName != null && $pubIdSettingValue != null && $pubIdSettingValue == DOI_EXPORT_STATUS_NOT_DEPOSITED)?' AND sss.setting_value IS NULL':'')
-				. (($pubIdSettingName != null && $pubIdSettingValue != null && $pubIdSettingValue != DOI_EXPORT_STATUS_NOT_DEPOSITED)?' AND sss.setting_value = ?':'')
+				. (($pubIdSettingName != null && $pubIdSettingValue != null && $pubIdSettingValue == EXPORT_STATUS_NOT_DEPOSITED)?' AND sss.setting_value IS NULL':'')
+				. (($pubIdSettingName != null && $pubIdSettingValue != null && $pubIdSettingValue != EXPORT_STATUS_NOT_DEPOSITED)?' AND sss.setting_value = ?':'')
 				. (($pubIdSettingName != null && is_null($pubIdSettingValue))?' AND (sss.setting_value IS NULL OR sss.setting_value = \'\')':'')
 			. ' ORDER BY ps.date_published DESC, s.submission_id DESC',
 			$params,
