+++ templates/controllers/grid/submissions/exportPublishedSubmissionsGridFilter.tpl
@@ -1,5 +1,5 @@
 {**
- * templates/controllers/grid/pubIds/pubIdExportSubmissionsGridFilter.tpl
+ * templates/controllers/grid/pubIds/exportPublishedSubmissionsGridFilter.tpl
  *
  * Copyright (c) 2014-2016 Simon Fraser University Library
  * Copyright (c) 2000-2016 John Willinsky
